requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
argcomplete>=3.0.0
